last update JJ

+++ src/start_cub.c
@@ -6,7 +6,7 @@
 /*   By: jjesberg <jjesberg@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/07 21:06:10 by jjesberg          #+#    #+#             */
-/*   Updated: 2023/01/14 14:55:41 by jjesberg         ###   ########.fr       */
+/*   Updated: 2023/01/14 15:23:29 by jjesberg         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 void	skyline(t_cub *cub)
 {
-	int32_t			i;
-	int32_t			j;
+	uint32_t			i;
+	uint32_t			j;
 
 	i = 0;
-	while (i < cub->s_height)
+	while (i < (uint32_t)cub->s_height)
 	{
 		j = 0;
-		while (j < cub->s_width)
+		while (j < (uint32_t)cub->s_width)
 		{
-			if (i < cub->s_height / 2)
+			if (i < (uint32_t)cub->s_height / 2)
 				mlx_put_pixel(cub->image, j, i, cub->d->c_colour);
 			else
 				mlx_put_pixel(cub->image, j, i, cub->d->f_colour);
